requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision
lightning>=2.0.0
transformers>=4.20.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Medical NLP Models
# Note: These will be downloaded automatically from Hugging Face
# - microsoft/biogpt
# - yikuan8/Clinical-Longformer  
# - emilyalsentzer/Bio_ClinicalBERT
# - medicalai/ClinicalBERT

# Traditional ML Models
catboost>=1.1.0
xgboost>=1.6.0

# Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Model Interpretability
shap>=0.41.0
transformers-interpret>=0.9.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
argparse

# Optional: For enhanced visualizations
imgkit  # Requires wkhtmltopdf system dependency

# Development (optional)
jupyter
ipython

# Note: For production use, consider pinning exact versions
# Example: torch==2.3.0 instead of torch>=2.0.0
